Move functions from SolrMarc -> MarcAdvancedTrait that support new ris fields
