+++ module/VuFind/src/VuFind/RecordDriver/Feature/MarcAdvancedTrait.php
@@ -35,6 +35,7 @@ namespace VuFind\RecordDriver\Feature;
 use VuFind\View\Helper\Root\RecordLinker;
 use VuFind\XSLT\Processor as XSLTProcessor;
 
+use function array_key_exists;
 use function count;
 use function in_array;
 use function is_array;
@@ -1169,5 +1170,68 @@ trait MarcAdvancedTrait
     public function getTextualHoldings()
     {
         return $this->getFieldArray('866');
+    }
+
+    /**
+     * Takes a Marc field that notes are stored in (ex: 950) and a list of
+     * sub fields (ex: ['a','b']) optionally as well as what indicator
+     * numbers and values to filter for and concatenates the subfields
+     * together and returns the fields back as an array
+     * (ex: ['subA subB subC', 'field2SubA field2SubB'])
+     *
+     * @param string $field    Marc field to search within
+     * @param ?array $subfield Sub-fields to return or empty for all
+     * @param array  $indData  Array containing the indicator number as the key
+     * and the value as an array of strings for the allowed indicator values
+     * ex: ['1' => ['1', '2'], '2' => ['']] would filter fields ind1 = 1 or 2 or ind2 = blank
+     *
+     * @return array The values within the subfields under the field
+     */
+    public function getMarcFieldWithInd(
+        string $field,
+        ?array $subfield = null,
+        array $indData = []
+    ) {
+        $vals = [];
+        $marc = $this->getMarcReader();
+        $marc_fields = $marc->getFields($field, $subfield);
+        foreach ($marc_fields as $marc_data) {
+            $field_vals = [];
+            // Check if that field has either indicator (MARC only has up to 2 indicators)
+            foreach (range(1, 2) as $indNum) {
+                if (array_key_exists($indNum, $indData)) {
+                    if (in_array(trim(($marc_data['i' . $indNum] ?? '')), $indData[$indNum])) {
+                        $subfields = $marc_data['subfields'];
+                        foreach ($subfields as $subfield) {
+                            $field_vals[] = $subfield['data'];
+                        }
+                    }
+                }
+            }
+            if (!empty($field_vals)) {
+                $vals[] = implode(' ', $field_vals);
+            }
+        }
+        return array_unique($vals);
+    }
+
+    /**
+     * Get the abstract and summary notes
+     *
+     * @return array Note fields from the MARC record
+     */
+    public function getAbstractAndSummaryNotes()
+    {
+        return $this->getMarcFieldWithInd('520', null, [1 => ['', '0', '2', '3', '8']]);
+    }
+
+    /**
+     * Get the location of other archival materials notes
+     *
+     * @return array Note fields from the MARC record
+     */
+    public function getLocationOfArchivalMaterialsNotes()
+    {
+        return $this->getMarcFieldWithInd('544', null, [1 => ['', '0']]);
     }
 }
