+++ module/VuFind/tests/unit-tests/src/VuFindTest/RecordDriver/Feature/MarcAdvancedTraitTest.php
@@ -204,4 +204,98 @@ class MarcAdvancedTraitTest extends \PHPUnit\Framework\TestCase
             $obj->getTitleSectionsAltScript()
         );
     }
+
+    /**
+     * Test getMarcFieldWithInd when a single indicator value is sent
+     *
+     * @return void
+     */
+    public function testGetMarcFieldWithIndOneValue(): void
+    {
+        $obj = $this->getMockDriverFromFixture('marc/marctraits.xml');
+
+        // Test when a single ind value is passed
+        $this->assertEquals(
+            ['upc'],
+            $obj->getMarcFieldWithInd('024', null, ['1' => ['1']])
+        );
+    }
+
+    /**
+     * Test getMarcFieldWithInd when multiple values for the indicator are sent
+     *
+     * @return void
+     */
+    public function testGetMarcFieldWithIndTwoValues(): void
+    {
+        $obj = $this->getMockDriverFromFixture('marc/marctraits.xml');
+
+        // Test when multiple values for the same ind are passed
+        $this->assertEquals(
+            ['upc', 'ismn'],
+            $obj->getMarcFieldWithInd('024', null, ['1' => ['1', '2']])
+        );
+    }
+
+    /**
+     * Test getMarcFieldWithInd when multiple indicators are requested
+     *
+     * @return void
+     */
+    public function testGetMarcFieldWithIndMultInds(): void
+    {
+        $obj = $this->getMockDriverFromFixture('marc/marctraits.xml');
+
+        // Test when different ind values are passed for each ind
+        $this->assertEquals(
+            ['ger', 'spa'],
+            $obj->getMarcFieldWithInd('041', null, ['1' => ['0'], '2' => ['7']])
+        );
+    }
+
+    /**
+     * Test getMarcFieldWithInd when no indicator filters are sent
+     *
+     * @return void
+     */
+    public function testGetMarcFieldWithIndNoValues(): void
+    {
+        $obj = $this->getMockDriverFromFixture('marc/marctraits.xml');
+
+        // Test when no indicator is passed
+        $this->assertEquals(
+            [],
+            $obj->getMarcFieldWithInd('024', null, [])
+        );
+    }
+
+    /**
+     * Test calling getAbstractAndSummaryNotes to get expected marc data
+     *
+     * @return void
+     */
+    public function testGetAbstractAndSummaryNotes(): void
+    {
+        $obj = $this->getMockDriverFromFixture('marc/marctraits.xml');
+
+        $this->assertEquals(
+            ['Summary. Expanded.'],
+            $obj->getAbstractAndSummaryNotes()
+        );
+    }
+
+    /**
+     * Test calling getLocationOfArchivalMaterialsNotes to get expected marc data
+     *
+     * @return void
+     */
+    public function testGetLocationOfArchivalMaterialsNotes(): void
+    {
+        $obj = $this->getMockDriverFromFixture('marc/marctraits.xml');
+
+        $this->assertEquals(
+            ['Location of archival materials'],
+            $obj->getLocationOfArchivalMaterialsNotes()
+        );
+    }
 }
